+++ src/webpack.js
@@ -37,8 +37,8 @@ module.exports = {
     gulp.task(taskName, [cleanUpTaskName], (done) => {
       const bundler = webpack(parameters.config);
       let counter = 0;
-      const cb = (err/*, stats*/) => {
-        //console.log(stats.toString(parameters.config.stats || {})); //eslint-disable-line no-console
+      const cb = (err /* , stats */) => {
+        // console.log(stats.toString(parameters.config.stats || {})); //eslint-disable-line no-console
 
         if (++counter === (parameters.config.length || 1)) {
           done(err);