+++ src/jscs.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const jscs = require('gulp-jscs');
+const path = require('path');
 
 /**
  * @module tasks/jscs
@@ -26,10 +27,39 @@ module.exports = {
    * );
    */
   register(gulp, taskName, parameters) {
-    gulp.task(taskName, () => {
-      return gulp.src(parameters.files)
-        .pipe(jscs({ configPath: parameters.configFile }))
+    const validate = (source, fix) => {
+      return gulp.src(source)
+        .pipe(jscs({
+          configPath: parameters.configFile,
+          fix: !!fix
+        }))
         .pipe(jscs.reporter());
+    };
+
+    gulp.task(taskName, () => validate(parameters.files));
+
+    gulp.task(taskName + ':fix', (done) => {
+      const filesByFolderPath = {};
+
+      gulp.src(parameters.files)
+        .on('error', err => done(err))
+        .on('data', (file) => {
+          const folderPath = path.dirname(file.path);
+          filesByFolderPath[folderPath] = filesByFolderPath[folderPath] || [];
+          filesByFolderPath[folderPath].push(file.path);
+        })
+        .on('end', () => {
+          Promise.all(
+            Object.keys(filesByFolderPath).map(folderPath => {
+              return new Promise((resolve, reject) => {
+                validate(filesByFolderPath[folderPath], true)
+                  .pipe(gulp.dest(folderPath))
+                  .on('error', err => reject(err))
+                  .on('end', () => resolve());
+              });
+            })
+          ).then(() => done(), err => done(err));
+        });
     });
   }
 };