+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Enhance jscs task (extra rules for jsdoc, fix reporter)
+- Be strict in ESLint rules (no warnings)
 
 ## 0.1.11 (2016-01-19)
 - Fix jest coverage report