+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Enhance jscs task (extra rules for jsdoc, fix reporter)
 - Be strict in ESLint rules (no warnings)
+- Introduce jscs:fix task
 
 ## 0.1.11 (2016-01-19)
 - Fix jest coverage report