+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Be strict in ESLint rules (no warnings)
 - Introduce jscs:fix task
 - Introduce eslint:fix task
+- Introduce retire task
 
 ## 0.1.11 (2016-01-19)
 - Fix jest coverage report